Update default version, build to reflect current package
The last packaged version for this repository that we deployed at
Betable was 1.0.3, with a package version of 1.0.3-betable1.

We update the version and build in a couple of ways:

- the version now starts with 1.0.4, the next build number

- instead of putting 'betable' in the build, add a '+betable' suffix
  to the version. This indicates that the code itself is specific to
  Betable.

- on Debian-based systems, append to the version a record of the OS
  version

Regarding that last point, some (Debian) repository managment systems
don't allow reusing the exact same package name for different
distributions. If the `/etc/lsb-release` script exists, use it to
construct a suffix of the form `"+${DISTRIB_ID}+${DISTRIB_RELEASE}"`.

The build just goes back to being an integer.

For example, on Ubuntu systems this should create versions

```
1.0.4+betable+ubuntu+14.04-1
1.0.4+betable+ubuntu+18.04-1
```